updating ctl script

+++ RHEL/6/input/fixes/bash/disable_ctrlaltdel_reboot.sh
@@ -1,5 +1,16 @@
-if [ -f /etc/init/control-alt-delete.conf ]; then
-	cp /etc/init/control-alt-delete.conf /etc/init/control-alt-delete.override
+# If system does not contain control-alt-delete.override,
+if [ ! -f /etc/init/control-alt-delete.override ]; then
+
+	# but does have control-alt-delete.conf file,
+	if [ -f /etc/init/control-alt-delete.conf ]; then
+
+		# then copy .conf to .override to maintain persistency
+		cp /etc/init/control-alt-delete.conf /etc/init/control-alt-delete.override
+	
+	fi
+
+	
 fi
 
-sed -i 's,^exec.*,exec /usr/bin/logger -p authpriv.notice -t init "Ctrl-Alt-Del was pressed and ignored",' /etc/init/control-alt-delete.override
+if 
+	sed -i 's,^exec.*,exec /usr/bin/logger -p authpriv.notice -t init "Ctrl-Alt-Del was pressed and ignored",' /etc/init/control-alt-delete.override
